+++ src/main/java/ec/com/jmgorduez/ACMEEmployeesPayment/domain/abstractions/IPayable.java
@@ -1,5 +1,5 @@
 package ec.com.jmgorduez.ACMEEmployeesPayment.domain.abstractions;
 
 public interface IPayable {
-    Integer payment();
+    Float payment();
 }
